Splitting up unitytrainers tests.
